Divide game tests into 2 collections
